tests/beamCG: fix bug in clamp boundary condition
In short, we are not allowed to prescribe the x component of the tangent, because this would impose a zero axial strain at the clamped end.

Longer story:
The tangent vector r' is a unit vector if and only if the axial strain is zero. Therefore imposing that the tangent vector is [1, 0, 0] implies that both 1) the tangent is horizontal, and 2) the strain at the clamped end is zero.
However, in general we want to only impose condition 1) as we do not know the strain at the clamped end a priori. Therefore imposing a horizontal tangent (e.g. aligned with the x-axis) is achieved by imposing that the y and z component of the tangent be fixed and equal to 0, whereas the x component of the tangent is free (zero Neumann boundary condition). See also Meier 2014, CMAME for more details.

Thanks @SaiKubairTUD for reporting this bug!