Modified penalty terms and parameters
Modified the quadratic penalty terms and the corresponding parameters. The penalty terms are modified such that the material, geometric and mesh information of the problem are taken into account. The units of the penalty terms are also consistent with the units of the remaining terms in the weak form. Currently, penalty parameters are just dimensionless numerical constants used in the penalty terms.