Modified the tensile DG/CZM residual and stiffness
Added two separate parameters as internal variables to indicate the damage status and to activate/deactivate DG/CZM interface force terms. Included axial jump value as a condition in the system residual function to test the possibility of recontact during loading. Added a condition in the system residual function to handle damage initiation possibility in case of very small separation. In the current implementation, the DG interface terms related to the forces and their tangents are deactivated upon damage initiation but the terms related to moments are kept active even after damage initiation. A better approach during a pure tensile damage would be to only deactivate the axial forces and add cohesive axial forces. The shear force and bending moment terms shouldn't be deactivated in that case.