Internal variables: Effective force and updates
Added effective force at damage initiation as an internal variable. This can be used in the TSL calculations to avoid convergence issues. The effective force at damage initiation is not used for now in the calculations. Feature to update the internal variables is added in the residual method. This feature should be used in the explicit analyses for updating the internal variables at every time step.