Internal Vars: Position jumps at damage initiation
Added position jumps at damage initiation as internal variables. These variables are updated as soon as damage is initiated at a particular interface. The syntax of the methods from the CohesiveInterfaceMaterial are also updated to match the latest implementation considering position jumps at damage initiation to compute the effective separation.