Modified the computation of system internal forces
Modified the internal force computation from using the elemental internal forces to computing the point-wise nodal forces directly. This is actually a major bug because the equivalence of elemental internal forces and point-wise nodal forces doesn't has to hold necessarily in dynamics. So the "nodal forces" are computed directly and assembled to form the system nodal forces both in the CG and DG discretization.