Removed either side shear, added constrained shear
Removed the previous separate shear force computation on either sides of the interface and adding them to the residual. The current implementation constrains shearing in the beam across the cohesive boundary during the fracture process through the DG shear flux and compatibility terms. These DG shear terms are removed altogether from the residual after complete fracture of the interface.