Constrained shear forces derivative coefficients
Implemented the constrained shear forces derivative coefficients. This function partially (some terms) computes the derivative of the constrained shear forces and can be used to compute the coefficients of the constrained shear forces derivatives that will be needed in the system stiffness. The remaining constrained shear forces derivative terms are assembled directly in the system.