Computing updated mass and complex Eigenvalues
Computing the mass matrix of the system including the rotational mass for evaluating the frequencies. Complex Eigenvalues are considered as a warning and ignored for now but these could lead to an instability in the system when excited. The governing mode's frequency is checked for a potential complex part and the analysis is terminated in that's true.