Post-fracture shear forces and delta_max
Added the correct value of shear forces after complete fracture. Modified the maximum effective separation after complete fracture to the maximum allowed separation (delta_c) at the interfaces. There is a bug in the earlier version of the code since it allows for forces and moments at the interfaces after complete fracture (during "closing in").