Static condensation bug in NewtonRaphson solver
Corrected the static condensation bug in the NewtonRaphson solver. The earlier version of the solver updated the Dirichlet solutions before the first iteration, and assembled & solved the system which could result in longer convergence durations. The current version updates the Dirichlet solution after the first iteration and the increment in the Dirichlet dofs is applied on the reduced system through static condensation (as internal forces in the right hand side) in the first iteration. The updates after the first iteration are carried out as usually without any forces resulting from static condensation.