System: initialize WeakForm based on FunctionSpace
Initialize the WeakForm based on the type of FunctionSpace input. For usual FunctionSpace, we initialize the beam KTLF weak form but for the EulerBernoulliFunctionSpace, the EulerBernoulliWeakForm is initialized. This is equivalent to choosing the type of element you need for solving the problem in beamit which depends on the type of FunctionSpace.