EBWeakForm: added compute system mass method
Added the compute system mass method in the Euler-Bernoulli weak form. This method is based on the latest implementation which includes axial dofs along with the bending dofs. Mass matrix for axial and bending contributions are computed separately (lumped if necessary) and added at the end to return the overall element mass.