EBFunctionSpace: added axial lifting shape functions
Added the axial lifting shape functions and an associated method for their computation in the EulerBernoulli function space. These lifting terms will be used in the DG derivative to encode the discontinuity information at the boundary. The DG derivative thus obtained will result in the flux and symmetrizing terms when used in the weak form.