EBWeakForm: method for element lifting forces
Added a method to compute the element lifting forces in the EulerBernoulli weak form. This method computes the variational lifting part of the internal forces in the discrete system of equations which comes out of the weak form when the DG derivative with lifting operator is substituted in the internal virtual work.