EBWeakForm: system residual with lifting terms
Added a method to compute the system residual of the DG EulerBernoulli weak form which takes into account the lifting term contributions which come along with the DG derivative that includes the lifting operator. This method also computes the boundary dof jumps which are needed for evaluating the lifting terms in the DG derivative and adds the element lifting forces at appropriate positions of the discrete internal force vector.