EBWeakForm: part of bend. lifting in lift. forces
Added the first part of the bending lifting operator contribution which comes from the DG double derivative of transverse displacement to the element lifting forces. These lifting forces will be assembled at the appropriate locations in the element residual to form the complete internal forces. However, it is observed that the numerical scheme is not completely stable with this implementation when the stabilization terms were switched off. Probably adding the second part of the bending lifting operator should correct this issue.