EBWeakForm: modified the lifting forces assembly
Modified the assembly of the lifting forces in the system residual. Now the lifting forces will only be assembled for a DG discretization. In the previous implementation, the element lifting forces were calculated for CG too which is incorrect since the derivatives of displacements over the elements without lifting contributions are nonzero.