EBWeakForm: added element extended lifting forces
Added the element extended lifting forces to the residual. These forces are a result of the mixed lifting variational term in the DG weak form of the bending equation. A general observation regarding these terms is that when they are added to the residual, the rotation related solution is losing a little bit of accuracy which should be further investigated. But in general, the overall compatibility of scheme has improved.