FunctionSpace: added an abstract FunctionSpace class & doc strings
Added an abstract FunctionSpace class to define the general structure of the FunctionSpace classes. The other more specialized classes now inherit from this base class which makes the structure of the FunctionSpace class more clearer and extendible. We also added doc strings for different methods to explain what they do and how they should be used.