WeakForm: running CZM calculations only when update_internal signal is received
We are running the CZM calculations only when the update_internal signal is received. This is not a good solution, as update_internal is not something that should be used for signaling CZM calculations. We need to think about some other fix for this in the future but let's keep it like this for now. The main idea here is to prevent fracture in, for example, implicit solvers by explicitly passing some signal to the weak form.