WeakForm: corrected bug in residual calculation of shear flexible geometrically beam
Corrected a bug in the residual calculation of shear flexible geometrically exact beam where the shape function terms was replaced by its derivative in one of the terms which resulted in a wrong residual calculation. We corrected it now and also consistently modified the stiffness calculation. In addition, some other minor changes corresponding to element force and moment calculations were also made in this commit.