SFGEWeakFormCG: added function to compute dof derivatives for easy extension to DG formulation
Added a function to compute the dof derivatives in the ShearFlexibleGeometricallyExactWeakFormCG class which enables an easier extension of the methods in the CG weak form to the ones in the DG weak form by just implementing a specialized derivative calculation. In addition, we also had to add element index to different methods since we will have to extract some elemental information during derivative calculation in the DG weak form.