SFGEWeakFormDG: corrected bug in interface residual assembly
Corrected a bug in the assembly of the interface residual terms in the ShearFlexibleGeometricallyExactWeakFormDG class where the subsequent indexing was creating a copy of the original array instead of modifying it in place. So it was not adding the interface residual terms to the corresponding entries in the global residual. The issue is now fixed and the interface residual terms are correctly added to the global residual.