WeakForm: added separate method to update the rotational solution
Added a separate method to update the rotational solution as we need to convert the additive update of the rotation which we implemented recently to multiplicative update first before proceeding with the internal variable updates. So we handle this in the weak form class itself which is passed down by the solver.