SFGEWeakFormDG: working lifting operator implementation
It seems like the lifting operator implementation is working correctly now at least for the case with one quad point. With more than one quad point, there is still some issue which we need to figure out. For now, let's run the convergence tests and see if the results are as expected.