lib: changed size of internal forces container so each element can write its own internal forces at nodes
Changed the size of the internal forces container in the System so that each element can write its own internal forces at the nodes. This makes the update of internal forces the same for both CG and DG discretizations. Also, internal forces and moments need not be continuous across elements in a CG discretization, since discontinuities can be present in the derivatives of the primary variables.